atualizando experimento
