+++ novaTentativa/app/src/main/java/com/example/bolsista/novatentativa/Resultado.java
@@ -1,41 +1,64 @@
 package com.example.bolsista.novatentativa;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.util.Log;
+import android.widget.Toast;
 
-import com.example.bolsista.novatentativa.adapters.ResultadoAdapter;
 import com.example.bolsista.novatentativa.arquitetura.Servidor;
+import com.example.bolsista.novatentativa.modelo.Experimento;
+import com.example.bolsista.novatentativa.viewsModels.ExperimentoViewModel;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 public class Resultado extends AppCompatActivity {
-    private RecyclerView desafiosRecycle;
-    private ResultadoAdapter adapter;
-
     private Context contextoAtivity;
+
+    //FireBase FireStore
+    FirebaseFirestore db = FirebaseFirestore.getInstance();
+    DocumentReference usuarioRef;
+    //FireBase autenth
+    FirebaseAuth usuario;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_resultado);
 
         inicializar();
-        if(Servidor.serverIdentificado) {
-            implementsRecycle();
-        }
+        updateExperimento();
     }
 
-    private void implementsRecycle() {
-        adapter = new ResultadoAdapter(contextoAtivity,Jogar.desafios);
-        desafiosRecycle.setAdapter(adapter);
-
-        LinearLayoutManager layout = new LinearLayoutManager(contextoAtivity,LinearLayoutManager.VERTICAL,false);
-        desafiosRecycle.setLayoutManager(layout);
+    private void updateExperimento() {
+        db.collection("experimentos")
+                .document(ExperimentoViewModel.experimento.getValue().getId())
+                .set(ExperimentoViewModel.experimento.getValue())
+                .addOnSuccessListener(new OnSuccessListener<Void>() {
+                    @Override
+                    public void onSuccess(Void aVoid) {
+                        Log.d("FireStore", "DocumentSnapshot successfully updated!");
+                        Toast.makeText(contextoAtivity, "Experimento Atualizado", Toast.LENGTH_SHORT).show();
+                    }
+                })
+                .addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Log.w("FireStore", "Error updating document", e);
+                    }
+                });
     }
 
     private void inicializar() {
-        desafiosRecycle = findViewById(R.id.desafiosRecycle);
         contextoAtivity = this;
+
+        usuario = FirebaseAuth.getInstance();
+        usuarioRef = db.collection("users").document(usuario.getCurrentUser().getUid());
     }
 }
