--- novaTentativa/app/src/main/res/layout/resultado_inflater.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:cardCornerRadius="5dp"
-    app:cardBackgroundColor="@color/cardview_light_background"
-    android:layout_margin="5dp"
-    android:focusable="true"
-    android:clickable="true"
-    android:foreground="?android:attr/selectableItemBackground"
-    >
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:gravity="center"
-        >
-        <ImageView
-            android:id="@+id/imgCorreta"
-            android:layout_weight="1"
-            android:layout_width="0dp"
-            android:layout_height="70sp"
-            android:src="@drawable/circulo"
-            >
-        </ImageView>
-
-        <ImageView
-            android:id="@+id/imgErrada"
-            android:layout_weight="1"
-            android:layout_width="0dp"
-            android:layout_height="70sp"
-            android:src="@drawable/triangulo"
-            >
-        </ImageView>
-
-        <TextView
-            android:id="@+id/qtdErros"
-            android:layout_weight="1"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:textColor="@color/black_de"
-            android:text="5"
-            android:gravity="center"
-            android:textSize="30sp"
-            >
-        </TextView>
-
-        <TextView
-            android:id="@+id/tempo"
-            android:layout_weight="1"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:textColor="@color/black_de"
-            android:text="8"
-            android:gravity="center"
-            android:textSize="30sp"
-            >
-        </TextView>
-    </LinearLayout>
-</androidx.cardview.widget.CardView>
